refactor : 채팅 기록api 연결, 코드 수정(#146)

+++ app/src/main/java/com/example/areumdap/UI/record/RecordFragment.kt
@@ -94,7 +94,7 @@ class RecordFragment : Fragment() {
 private fun UserChatThread.toRecordItem(): RecordItem {
     val parsed = summary.parseSummaryJson()
     return RecordItem(
-        id = threadId.toInt(),
+        id = threadId,
         category = tag.toCategory(),
         title = parsed.title.ifBlank { content.trim() },
         summary = parsed.summary,
